+++ docs/content/release-notes.md
@@ -1,4 +1,15 @@
+#### 0.3.0 - January 16, 2016
+* breaking change
+* VerbEx now fully composable
+* Match' composable System.Text.RegularExpressions.Match type
+* Group' composable System.Text.RegularExpressions.Group type
+* System.Text.RegularExpressions special collections returned as arrays
+* VerbEx timeout constructor added
+* added exists, fold, foldBack, iter, and map for VerbalExpression module
+* eliminated underscores in unicode names
+
 #### 0.2.0 - January 10, 2016
+* breaking change
 * put UniCodeGeneralCategory and SupportedNamedBlock under the namespace
 
 #### 0.1.0 - January 9, 2016